refactor(core): extract auth OIDC plumbing out of mod.rs into oidc.rs
Every other module in redisctl-core keeps a thin facade mod.rs (config,
cloud, enterprise are all mod/use only); auth/mod.rs was the outlier,
carrying TokenSet, AuthError, and the shared token-endpoint helpers.

Move those into a private `oidc` submodule and re-export them, so
auth/mod.rs is a facade like its siblings. Pure move: the flow/client
files are unchanged (they still reference the helpers via `super::`),
no behaviour change, tests untouched.
